ml-service/requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
requests>=2.31.0
httpx>=0.25.2
python-multipart>=0.0.6
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
loguru>=0.7.2

# Data & ML core
pandas>=2.1.3
numpy>=1.26.2,<2.0.0
scikit-learn>=1.3.2
joblib>=1.3.2

# Optional: ML tools (may require C compilation)
# These are installed separately in CI to avoid blocking core deps
matplotlib>=3.7.0
mlflow>=2.8.1
shap>=0.43.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0
